requirements.txt
colorama==0.4.6
fastapi==0.95.1
langchain==0.0.157
llama_index==0.6.0
llama-cpp-python==0.1.44
mistune==2.0.5
nltk==3.8.1
openai==0.27.6
pexpect==4.8.0
Pillow==9.5.0
protobuf==3.20.3
Pygments==2.15.1
python-dotenv==1.0.0
rake_nltk==1.0.6
requests==2.29.0
rich==13.3.5
sentencepiece==0.1.98
SQLAlchemy==2.0.12
torch==1.13.1
transformers==4.28.1
